refactor(foundation-ui): move DCOSStore and ResourceTableUtil from founcdation
because they don't belong there
