+++ foundation-ui/navigation/NavigationService.js
@@ -1,6 +1,5 @@
 import {EventEmitter} from 'events';
 import {NAVIGATION_CHANGE} from './EventTypes';
-import Config from '../../src/js/config/Config';
 
 class NavigationService extends EventEmitter {
 
@@ -92,7 +91,7 @@ class NavigationService extends EventEmitter {
     });
 
     if (existingElement) {
-      if (Config.environment === 'development') {
+      if (process.env.NODE_ENV === 'development') {
         console.warn(`Primary nav with the path ${path} already exists!`);
       }
 
@@ -137,7 +136,7 @@ class NavigationService extends EventEmitter {
     });
 
     if (existingElement) {
-      if (Config.environment === 'development') {
+      if (process.env.NODE_ENV === 'development') {
         console.warn(`
           Secondary nav with the path ${parentPath}/${path} already exists!
         `);
