+++ foundation-ui/package.json
@@ -1,0 +1,13 @@
+{
+  "name": "foundation-ui",
+  "main": "index.js",
+  "version": "0.1.0",
+  "dependencies": {
+    "react": "0.14.7",
+    "react-router": "2.8.1"
+  },
+  "devDependencies": {
+    "react-addons-test-utils": "0.14.7",
+    "jest-cli": "16.0.2"
+  }
+}
