+++ plugins/nodes/src/js/pages/nodes/NodeDetailTaskTab.js
@@ -1,10 +1,10 @@
-import {DCOSStore} from 'foundation-ui';
 import React from 'react';
 import {routerShape} from 'react-router';
 
 import {
   DCOS_CHANGE
 } from '../../../../../../src/js/constants/EventTypes';
+import DCOSStore from '../../../../../../src/js/stores/DCOSStore';
 
 import MesosStateStore from '../../../../../../src/js/stores/MesosStateStore';
 import TasksContainer from '../../../../../services/src/js/containers/tasks/TasksContainer';
