+++ plugins/services/src/js/components/MarathonTaskDetailsList.js
@@ -1,6 +1,6 @@
-import {DCOSStore} from 'foundation-ui';
 import React from 'react';
 
+import DCOSStore from '../../../../../src/js/stores/DCOSStore';
 import ConfigurationMapHeading from '../../../../../src/js/components/ConfigurationMapHeading';
 import ConfigurationMapLabel from '../../../../../src/js/components/ConfigurationMapLabel';
 import ConfigurationMapRow from '../../../../../src/js/components/ConfigurationMapRow';
