+++ plugins/services/src/js/pages/__tests__/DeploymentsTab-test.js
@@ -14,7 +14,7 @@ const React = require('react');
 /* eslint-enable no-unused-vars */
 const ReactDOM = require('react-dom');
 
-const DCOSStore = require('foundation-ui').DCOSStore;
+const DCOSStore = require('../../../../../../src/js/stores/DCOSStore');
 const Deployment = require('../../structs/Deployment');
 const DeploymentsTab = require('../services/DeploymentsTab');
 const DeploymentsList = require('../../structs/DeploymentsList');
