+++ src/js/pages/DashboardPage.js
@@ -1,4 +1,3 @@
-import {DCOSStore} from 'foundation-ui';
 import {routerShape, Link} from 'react-router';
 import React from 'react';
 import {StoreMixin} from 'mesosphere-shared-reactjs';
@@ -7,6 +6,7 @@ import Breadcrumb from '../components/Breadcrumb';
 import BreadcrumbTextContent from '../components/BreadcrumbTextContent';
 import ComponentList from '../components/ComponentList';
 import Config from '../config/Config';
+import DCOSStore from '../stores/DCOSStore';
 import HealthSorting from '../../../plugins/services/src/js/constants/HealthSorting';
 import HostTimeSeriesChart from '../components/charts/HostTimeSeriesChart';
 import Icon from '../components/Icon';
