+++ src/js/pages/jobs/JobsTab.js
@@ -1,4 +1,3 @@
-import {DCOSStore} from 'foundation-ui';
 import mixin from 'reactjs-mixin';
 import React from 'react';
 import {routerShape} from 'react-router';
@@ -7,6 +6,7 @@ import {StoreMixin} from 'mesosphere-shared-reactjs';
 
 import AlertPanel from '../../components/AlertPanel';
 import AlertPanelHeader from '../../components/AlertPanelHeader';
+import DCOSStore from '../../stores/DCOSStore';
 import FilterBar from '../../components/FilterBar';
 import FilterHeadline from '../../components/FilterHeadline';
 import JobsBreadcrumbs from '../../components/breadcrumbs/JobsBreadcrumbs';
