+++ src/js/plugin-bridge/Loader.js
@@ -16,9 +16,8 @@ const requireIcons = require.context('../components/icons', false);
 const requireModals = require.context('../components/modals', false);
 const requireForm = require.context('../components/form', false);
 // Foundation
-const requireRouting = require.context('../../../foundation-ui/routing', false);
-const requireNavigation = require.context('../../../foundation-ui/navigation', false);
-const requireFoundationUtils = require.context('../../../foundation-ui/utils', false);
+const requireRouting = require.context('../../../foundation-ui/src/routing', false);
+const requireNavigation = require.context('../../../foundation-ui/src/navigation', false);
 const requireFoundation = require.context('../../../foundation-ui', false);
 let requireExternalPlugin = function () {
   return {};
@@ -103,8 +102,6 @@ function requireModule(dir, name) {
       return requireEvents(path);
     case 'routing':
       return requireRouting(path);
-    case 'foundation-utils':
-      return requireFoundationUtils(path);
     case 'foundation-ui':
       return requireFoundation(path);
     case 'systemPages':
