+++ src/js/plugin-bridge/PluginModules.js
@@ -10,6 +10,7 @@ module.exports = {
     SidebarActions: 'SidebarActions'
   },
   stores: {
+    DCOSStore: 'DCOSStore',
     AuthStore: 'AuthStore',
     BaseStore: 'BaseStore',
     ConfigStore: 'ConfigStore',
@@ -37,9 +38,10 @@ module.exports = {
     FormUtil: 'FormUtil',
     LocalStorageUtil: 'LocalStorageUtil',
     Maths: 'Maths',
-    RouterUtil: 'RouterUtil',
-    ReducerUtil: 'ReducerUtil',
     ParserUtil: 'ParserUtil',
+    ReducerUtil: 'ReducerUtil',
+    ResourceTableUtil: 'ResourceTableUtil',
+    RouterUtil: 'RouterUtil',
     StringUtil: 'StringUtil',
     TableUtil: 'TableUtil',
     TabsUtil: 'TabsUtil',
@@ -119,9 +121,6 @@ module.exports = {
   },
   'foundation-ui': {
     'foundation-ui': 'index'
-  },
-  'foundation-utils': {
-    ResourceTableUtil: 'ResourceTableUtil'
   },
   internalPlugin: {
     services: 'services/index'
