+++ webpack/webpack.config.babel.js
@@ -163,8 +163,7 @@ module.exports = {
       PluginSDK: absPath('src/js/plugin-bridge/PluginSDK'),
       PluginTestUtils: absPath('src/js/plugin-bridge/PluginTestUtils'),
       EXTERNAL_PLUGINS: externalPluginsDir,
-      PLUGINS: absPath('plugins'),
-      'foundation-ui': absPath('foundation-ui')
+      PLUGINS: absPath('plugins')
     },
     extensions: ['', '.js', '.less', '.css'],
     root: [absPath(), absPath('node_modules')]
